Avoid confusion between profiles.py and profiles directory.
Moved all code from ``profiles.py`` to ``setuphandlers.py`` to avoid confusion with ``profiles`` directory.
This avoids a warning at startup:

```
ImportWarning: Not importing directory '.../plone/app/iterate/profiles': missing __init__.py
```
