+++ news/3130.bugfix
@@ -1,2 +1,3 @@
 Fixed deprecation warning for ObjectEvent.
+Moved all code from ``profiles.py`` to ``setuphandlers.py`` to avoid confusion with ``profiles`` directory.
 [maurits]
